FIX: Special chars are removed and %20 will be filtered out after API call is done
Now every special character that ADO does not allow will be removed from the sprint name.
Every space from the project name will be replaced with a %20 so the URL can be read if we do a rest API call.
After we did the call the project name with spaces (%20) will be replaces with an actual space.